add evaluation of safety logs and support for multiple episodes

+++ simulation-ros/src/turtlebot2i/turtlebot2i_edge/scripts/evaluate.py
@@ -6,21 +6,13 @@ from __future__ import division
 import argparse
 import os
 import shutil
-import copy
 import numpy as np
 import matplotlib.pyplot as plt
 import seaborn as sns
 from turtlebot2i_edge import TaskOffloadingEnv
 
 
-def preprocess(logs):
-    # flatten
-    logs_ = copy.deepcopy(logs)
-    logs = {k: [] for k in logs_[0].keys()}
-    for episode_logs in logs_:
-        for k, v in episode_logs.items():
-            logs[k] += v
-
+def preprocess_offloading(logs):
     # boolean -> 0/1
     for k in ['published', 'edge_model']:
         logs[k] = [1 if bool_ else 0 for bool_ in logs[k]]
@@ -33,6 +25,16 @@ def preprocess(logs):
     for k, v in logs.items():
         if k != 'published':
             logs[k] = logs[k][idx_published]
+
+    return logs
+
+
+def preprocess_safety(logs):
+    # add risk x speed
+    logs['risk_value_x_speed'] = [risk_value * speed for risk_value, speed in zip(logs['risk_value'], logs['speed'])]
+
+    # list -> numpy array
+    logs = {k: np.array(v) for k, v in logs.items()}
 
     return logs
 
@@ -99,7 +101,8 @@ def plot_2d_density(logs, metric_x, metric_y, output_dir=None, filename=None):
 
 def get_command_line_arguments():
     parser = argparse.ArgumentParser(description='Evaluate task offloading agent.')
-    parser.add_argument('logs', type=str, help='path to the logs')
+    parser.add_argument('logs_offloading', type=str, help='path to the offloading logs')
+    parser.add_argument('logs_safety', type=str, help='path to the safety logs')
     parser.add_argument('--output', type=str, help='path where to store results')
     return parser.parse_args()
 
@@ -109,35 +112,64 @@ def main():
     if args.output is not None and os.path.isfile(args.output):
         raise ValueError('Invalid output path, it is a file')
     if args.output is not None:
-        shutil.rmtree(args.output)
+        if os.path.exists(args.output):
+            shutil.rmtree(args.output)
         os.mkdir(args.output)
 
-    logs = np.load(args.logs, allow_pickle=True)['logs']
-    logs = preprocess(logs)
+    logs_offloading = np.load(args.logs_offloading, allow_pickle=True)['logs']
+    logs_safety = np.load(args.logs_safety, allow_pickle=True)['logs']
+    assert(len(logs_offloading) == len(logs_safety))
+    n_episodes = len(logs_offloading)
 
-    for metric in ['reward', 'latency', 'energy']:
-        metric_mean, metric_std, metric_min, metric_max = get_metric_per_step(logs, metric)
-        output = '%s per step: %.2f +- %.2f [%.2f, %.2f]' % (metric, metric_mean, metric_std, metric_min, metric_max)
-        print_output(output, output_dir=args.output)
+    for episode in range(n_episodes):
+        episode_logs_offloading = preprocess_offloading(logs_offloading[episode])
+        episode_logs_safety = preprocess_safety(logs_safety[episode])
 
-    for metric in ['published', 'edge_model']:
-        metric_perc = get_percentage_metric(logs, 'published')
-        output = '%s: %.2f%%' % (metric, metric_perc)
-        print_output(output, output_dir=args.output)
+        print_output('===========', output_dir=args.output)
+        print_output('episode: %d' % (episode+1), output_dir=args.output)
+        print_output('===========', output_dir=args.output)
 
-    plot_action_pie(logs, output_dir=args.output)
-    plot_2d_density(logs, 'risk_value', 'latency', output_dir=args.output)
-    plot_2d_density(logs, 'risk_value', 'edge_model', output_dir=args.output)
+        for metric in ['reward', 'latency', 'energy']:
+            metric_mean, metric_std, metric_min, metric_max = get_metric_per_step(episode_logs_offloading, metric)
+            output = '%s per step: %.2f +- %.2f (min=%.2f, max=%.2f)' % \
+                     (metric, metric_mean, metric_std, metric_min, metric_max)
+            print_output(output, output_dir=args.output)
 
-    filtered_logs = filter_logs_by_action(logs, [2])
-    if filtered_logs is not None:
-        filename = 'risk_value_temporal_coherence_use_last.jpg'
-        plot_2d_density(filtered_logs, 'risk_value', 'temporal_coherence', output_dir=args.output, filename=filename)
+        for metric in ['published', 'edge_model']:
+            metric_perc = get_percentage_metric(episode_logs_offloading, 'published')
+            output = '%s: %.2f%%' % (metric, metric_perc)
+            print_output(output, output_dir=args.output)
 
-    filtered_logs = filter_logs_by_action(logs, [0, 1])
-    if filtered_logs is not None:
-        filename = 'risk_value_temporal_coherence_compute.jpg'
-        plot_2d_density(filtered_logs, 'risk_value', 'temporal_coherence', output_dir=args.output, filename=filename)
+        for metric in ['risk_value', 'risk_value_x_speed']:
+            metric_mean, metric_std, metric_min, metric_max = get_metric_per_step(episode_logs_safety, metric)
+            output = '%s per step: %.2f +- %.2f (min=%.2f, max=%.2f)' % \
+                     (metric, metric_mean, metric_std, metric_min, metric_max)
+            print_output(output, output_dir=args.output)
+
+        for metric in ['in_critical_zone', 'in_warning_zone', 'in_safe_zone', 'collisions', 'mean_distance',
+                       'mean_min_distance', 'min_distance']:
+            output = '%s: %f' % (metric, episode_logs_safety[metric])
+            print_output(output, output_dir=args.output)
+
+        plot_dir = os.path.join(args.output, 'episode_%d' % (episode+1))
+        os.mkdir(plot_dir)
+
+        plot_action_pie(episode_logs_offloading, output_dir=plot_dir)
+        plot_2d_density(episode_logs_offloading, 'risk_value', 'latency', output_dir=plot_dir)
+        plot_2d_density(episode_logs_offloading, 'risk_value', 'edge_model', output_dir=plot_dir)
+
+        filtered_logs = filter_logs_by_action(episode_logs_offloading, [2])
+        if filtered_logs is not None:
+            filename = 'risk_value_temporal_coherence_use_last.jpg'
+            plot_2d_density(filtered_logs, 'risk_value', 'temporal_coherence', output_dir=plot_dir, filename=filename)
+
+        filtered_logs = filter_logs_by_action(episode_logs_offloading, [0, 1])
+        if filtered_logs is not None:
+            filename = 'risk_value_temporal_coherence_compute.jpg'
+            plot_2d_density(filtered_logs, 'risk_value', 'temporal_coherence', output_dir=plot_dir, filename=filename)
+
+        if episode < len(logs_offloading)-1:
+            print_output('', output_dir=args.output)
 
 
 if __name__ == '__main__':
